+++ henkan.py
@@ -1,6 +1,10 @@
 import subprocess
 import sys
+import os
 input_file_0=open("result/"+sys.argv[1],'r')
+if os.path.exists("result/"+sys.argv[1]) != 1:
+    sys.exit(0)
+
 cmd = "touch "+"result/a-"+sys.argv[1]
 subprocess.run(cmd, shell=True)
 input_file_1=open("result/a-"+sys.argv[1],'w')
